# Micro optimise Symbol#fullName