@retronym 2016-02-09
This is the profile that led me in this direction.

![image](https://cloud.githubusercontent.com/assets/65551/12916087/85bef9f8-cf7b-11e5-8f25-ab954ef831c0.png)

![image](https://cloud.githubusercontent.com/assets/65551/12916120/ad0ffd36-cf7b-11e5-9c90-ec76e805c27d.png)

SBT uses `fullName` as well:

![image](https://cloud.githubusercontent.com/assets/65551/12916155/f9e7b98c-cf7b-11e5-9dbd-f3f861f495a9.png)